requirements.txt
langchain
langchain-anthropic
chromadb
psycopg2-binary
sqlalchemy
pandas
gspread
oauth2client
python-dotenv
langdetect
deep-translator
vaderSentiment
pyspellchecker
sentence-transformers
hdbscan
numpy
scikit-learn
ollama